ai-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
scikit-learn==1.3.2
nltk==3.8.1
numpy==1.26.2
pandas==2.1.3
pymongo==4.6.0
motor==3.3.2
python-dotenv==1.0.0
pydantic==2.5.0
joblib==1.3.2
httpx==0.25.2
